fix(auth): Move CapabilityManager into CapabilityService
